fixed test folder's issue
