+++ src/__test__/home.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import render from 'react-test-renderer';
-import Quote from '../../components/quotes';
+import Home from '../components/home';
 
-describe('Quote', () => {
-  const page = render.create(<Quote />).toJSON();
+describe('Home', () => {
+  const page = render.create(<Home />).toJSON();
   test('should render without crashing', () => {
     expect(page).toMatchSnapshot();
   });
