+++ src/__test__/quotes.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import render from 'react-test-renderer';
-import Display from '../../components/Display';
+import Quote from '../components/quotes';
 
-describe('Display', () => {
-  const page = render.create(<Display />);
+describe('Quote', () => {
+  const page = render.create(<Quote />).toJSON();
   test('should render without crashing', () => {
     expect(page).toMatchSnapshot();
   });
